--- a/ultralytics/cfg/__init__.py
+++ b/ultralytics/cfg/__init__.py
@@ -82,11 +82,9 @@ CLI_HELP_MSG = f"""
 
     5. Explore your datasets using semantic search and SQL with a simple GUI powered by Ultralytics Explorer API
         yolo explorer data=data.yaml model=yolo11n.pt
-        
 
     6. Streamlit real-time webcam inference GUI
         yolo streamlit-predict
-
 
     7. Run special commands:
         yolo help