--- a/ultralytics/cfg/solutions/default.yaml
+++ b/ultralytics/cfg/solutions/default.yaml
@@ -15,4 +15,4 @@ up_angle: 145.0 # Workouts up_angle for counts, 145.0 is default value
 down_angle: 90 # Workouts down_angle for counts, 90 is default value
 kpts: [6, 8, 10] # Keypoints for workouts monitoring
 
-colormap: COLORMAP_PARULA # Colormap for heatmap
+colormap:  # Colormap for heatmap