--- a/ultralytics/solutions/heatmap.py
+++ b/ultralytics/solutions/heatmap.py
@@ -1,249 +1,91 @@
 # Ultralytics YOLO 🚀, AGPL-3.0 license
 
-from collections import defaultdict
-
 import cv2
 import numpy as np
+from astroid.nodes.scoped_nodes.scoped_nodes import objects
 
-from ultralytics.utils.checks import check_imshow, check_requirements
 from ultralytics.utils.plotting import Annotator
-
-check_requirements("shapely>=2.0.0")
-
-from shapely.geometry import LineString, Point, Polygon
+from ultralytics.solutions.solutions import BaseSolution  # Import a parent class
 
 
-class Heatmap:
+class Heatmap(BaseSolution):
     """A class to draw heatmaps in real-time video stream based on their tracks."""
 
-    def __init__(
-        self,
-        names,
-        colormap=cv2.COLORMAP_JET,
-        view_img=False,
-        view_in_counts=True,
-        view_out_counts=True,
-        count_reg_pts=None,
-        count_txt_color=(0, 0, 0),
-        count_bg_color=(255, 255, 255),
-        count_reg_color=(255, 0, 255),
-        region_thickness=5,
-        line_dist_thresh=15,
-        line_thickness=2,
-        shape="circle",
-    ):
-        """Initializes the heatmap class with default values for Visual, Image, track, count and heatmap parameters."""
-        # Visual information
-        self.annotator = None
-        self.view_img = view_img
-        self.shape = shape
+    def __init__(self, **kwargs):
+        """Initializes function for heatmap class with default values."""
+        super().__init__(**kwargs)
 
-        self.initialized = False
-        self.names = names  # Classes names
+        if self.region is not None:
+            self.initialize_region()
+        self.initialized = False  # bool variable for heatmap initialization
 
-        # Image information
-        self.im0 = None
-        self.tf = line_thickness
-        self.view_in_counts = view_in_counts
-        self.view_out_counts = view_out_counts
+        self.colormap = cv2.COLORMAP_PARULA # Store colormap for heatmap
 
-        # Heatmap colormap and heatmap np array
-        self.colormap = colormap
-        self.heatmap = None
-
-        # Predict/track information
-        self.boxes = []
-        self.track_ids = []
-        self.clss = []
-        self.track_history = defaultdict(list)
-
-        # Region & Line Information
-        self.counting_region = None
-        self.line_dist_thresh = line_dist_thresh
-        self.region_thickness = region_thickness
-        self.region_color = count_reg_color
-
-        # Object Counting Information
-        self.in_counts = 0
-        self.out_counts = 0
-        self.count_ids = []
-        self.class_wise_count = {}
-        self.count_txt_color = count_txt_color
-        self.count_bg_color = count_bg_color
-        self.cls_txtdisplay_gap = 50
-
-        # Check if environment supports imshow
-        self.env_check = check_imshow(warn=True)
-
-        # Region and line selection
-        self.count_reg_pts = count_reg_pts
-        print(self.count_reg_pts)
-        if self.count_reg_pts is not None:
-            if len(self.count_reg_pts) == 2:
-                print("Line Counter Initiated.")
-                self.counting_region = LineString(self.count_reg_pts)
-            elif len(self.count_reg_pts) >= 3:
-                print("Polygon Counter Initiated.")
-                self.counting_region = Polygon(self.count_reg_pts)
-            else:
-                print("Invalid Region points provided, region_points must be 2 for lines or >= 3 for polygons.")
-                print("Using Line Counter Now")
-                self.counting_region = LineString(self.count_reg_pts)
-
-        # Shape of heatmap, if not selected
-        if self.shape not in {"circle", "rect"}:
-            print("Unknown shape value provided, 'circle' & 'rect' supported")
-            print("Using Circular shape now")
-            self.shape = "circle"
-
-    def extract_results(self, tracks):
+    def heatmap_effect(self, box):
         """
-        Extracts results from the provided data.
+        Efficient calculation of heatmap area and effect location for applying colormap.
 
         Args:
-            tracks (list): List of tracks obtained from the object tracking process.
+            box (list): Bounding Box coordinates data [x0, y0, x1, y1]
         """
-        if tracks[0].boxes.id is not None:
-            self.boxes = tracks[0].boxes.xyxy.cpu()
-            self.clss = tracks[0].boxes.cls.tolist()
-            self.track_ids = tracks[0].boxes.id.int().tolist()
+        x0, y0, x1, y1 = map(int, box)
+        radius_squared = (min(x1 - x0, y1 - y0) // 2) ** 2
 
-    def generate_heatmap(self, im0, tracks):
+        # Create a meshgrid with region of interest (ROI) for vectorized distance calculations
+        xv, yv = np.meshgrid(np.arange(x0, x1), np.arange(y0, y1))
+
+        # Calculate squared distances from the center
+        dist_squared = (xv - ((x0 + x1) // 2)) ** 2 + (yv - ((y0 + y1) // 2)) ** 2
+
+        # Create a mask of points within the radius
+        within_radius = dist_squared <= radius_squared
+
+        # Update only the values within the bounding box in a single vectorized operation
+        self.heatmap[y0:y1, x0:x1][within_radius] += 2
+
+    def generate_heatmap(self, im0):
         """
-        Generate heatmap based on tracking data.
+        Generate heatmap for each frame using Ultralytics.
 
         Args:
-            im0 (nd array): Image
-            tracks (list): List of tracks obtained from the object tracking process.
+            im0 (ndarray): Input image array for processing
+        Returns:
+            im0 (ndarray): Processed image for further usage
         """
-        self.im0 = im0
+        self.heatmap = np.zeros_like(im0, dtype=np.float32) * 0.99 if not self.initialized else self.heatmap
+        self.initialized = True  # Initialize heatmap only once
 
-        # Initialize heatmap only once
-        if not self.initialized:
-            self.heatmap = np.zeros((int(self.im0.shape[0]), int(self.im0.shape[1])), dtype=np.float32)
-            self.initialized = True
+        self.annotator = Annotator(im0, line_width=self.line_width)  # Initialize annotator
+        self.extract_tracks(im0)  # Extract tracks
 
-        self.heatmap *= 0.99  # decay factor
+        # Iterate over bounding boxes, track ids and classes index
+        if self.track_data is not None and self.track_data.id is not None:
+            for box, track_id, cls in zip(self.boxes, self.track_ids, self.clss):
+                # Draw bounding box and counting region
+                self.heatmap_effect(box)
 
-        self.extract_results(tracks)
-        self.annotator = Annotator(self.im0, self.tf, None)
+                if self.region is not None:
+                    self.annotator.draw_region(reg_pts=self.region, color=(104, 0, 123), thickness=self.line_width * 2)
+                    self.store_tracking_history(track_id, box)  # Store track history
+                    self.store_classwise_counts(cls)  # store classwise counts in dict
 
-        if self.track_ids:
-            # Draw counting region
-            if self.count_reg_pts is not None:
-                self.annotator.draw_region(
-                    reg_pts=self.count_reg_pts, color=self.region_color, thickness=self.region_thickness
-                )
+                    # Store tracking previous position and perform object counting
+                    prev_position = self.track_history[track_id][-2] if len(self.track_history[track_id]) > 1 else None
+                    self.count_objects(self.track_line, box, track_id, prev_position, cls)  # Perform object counting
 
-            for box, cls, track_id in zip(self.boxes, self.clss, self.track_ids):
-                # Store class info
-                if self.names[cls] not in self.class_wise_count:
-                    self.class_wise_count[self.names[cls]] = {"IN": 0, "OUT": 0}
-
-                if self.shape == "circle":
-                    center = (int((box[0] + box[2]) // 2), int((box[1] + box[3]) // 2))
-                    radius = min(int(box[2]) - int(box[0]), int(box[3]) - int(box[1])) // 2
-
-                    y, x = np.ogrid[0 : self.heatmap.shape[0], 0 : self.heatmap.shape[1]]
-                    mask = (x - center[0]) ** 2 + (y - center[1]) ** 2 <= radius**2
-
-                    self.heatmap[int(box[1]) : int(box[3]), int(box[0]) : int(box[2])] += (
-                        2 * mask[int(box[1]) : int(box[3]), int(box[0]) : int(box[2])]
-                    )
-
-                else:
-                    self.heatmap[int(box[1]) : int(box[3]), int(box[0]) : int(box[2])] += 2
-
-                # Store tracking hist
-                track_line = self.track_history[track_id]
-                track_line.append((float((box[0] + box[2]) / 2), float((box[1] + box[3]) / 2)))
-                if len(track_line) > 30:
-                    track_line.pop(0)
-
-                prev_position = self.track_history[track_id][-2] if len(self.track_history[track_id]) > 1 else None
-
-                if self.count_reg_pts is not None:
-                    # Count objects in any polygon
-                    if len(self.count_reg_pts) >= 3:
-                        is_inside = self.counting_region.contains(Point(track_line[-1]))
-
-                        if prev_position is not None and is_inside and track_id not in self.count_ids:
-                            self.count_ids.append(track_id)
-
-                            if (box[0] - prev_position[0]) * (self.counting_region.centroid.x - prev_position[0]) > 0:
-                                self.in_counts += 1
-                                self.class_wise_count[self.names[cls]]["IN"] += 1
-                            else:
-                                self.out_counts += 1
-                                self.class_wise_count[self.names[cls]]["OUT"] += 1
-
-                    # Count objects using line
-                    elif len(self.count_reg_pts) == 2:
-                        if prev_position is not None and track_id not in self.count_ids:
-                            distance = Point(track_line[-1]).distance(self.counting_region)
-                            if distance < self.line_dist_thresh and track_id not in self.count_ids:
-                                self.count_ids.append(track_id)
-
-                                if (box[0] - prev_position[0]) * (
-                                    self.counting_region.centroid.x - prev_position[0]
-                                ) > 0:
-                                    self.in_counts += 1
-                                    self.class_wise_count[self.names[cls]]["IN"] += 1
-                                else:
-                                    self.out_counts += 1
-                                    self.class_wise_count[self.names[cls]]["OUT"] += 1
-
-        else:
-            for box, cls in zip(self.boxes, self.clss):
-                if self.shape == "circle":
-                    center = (int((box[0] + box[2]) // 2), int((box[1] + box[3]) // 2))
-                    radius = min(int(box[2]) - int(box[0]), int(box[3]) - int(box[1])) // 2
-
-                    y, x = np.ogrid[0 : self.heatmap.shape[0], 0 : self.heatmap.shape[1]]
-                    mask = (x - center[0]) ** 2 + (y - center[1]) ** 2 <= radius**2
-
-                    self.heatmap[int(box[1]) : int(box[3]), int(box[0]) : int(box[2])] += (
-                        2 * mask[int(box[1]) : int(box[3]), int(box[0]) : int(box[2])]
-                    )
-
-                else:
-                    self.heatmap[int(box[1]) : int(box[3]), int(box[0]) : int(box[2])] += 2
-
-        if self.count_reg_pts is not None:
-            labels_dict = {}
-
-            for key, value in self.class_wise_count.items():
-                if value["IN"] != 0 or value["OUT"] != 0:
-                    if not self.view_in_counts and not self.view_out_counts:
-                        continue
-                    elif not self.view_in_counts:
-                        labels_dict[str.capitalize(key)] = f"OUT {value['OUT']}"
-                    elif not self.view_out_counts:
-                        labels_dict[str.capitalize(key)] = f"IN {value['IN']}"
-                    else:
-                        labels_dict[str.capitalize(key)] = f"IN {value['IN']} OUT {value['OUT']}"
-
-            if labels_dict is not None:
-                self.annotator.display_analytics(self.im0, labels_dict, self.count_txt_color, self.count_bg_color, 10)
+        self.display_counts(im0) if self.region is not None else None  # Display the counts on the frame
 
         # Normalize, apply colormap to heatmap and combine with original image
-        heatmap_normalized = cv2.normalize(self.heatmap, None, 0, 255, cv2.NORM_MINMAX)
-        heatmap_colored = cv2.applyColorMap(heatmap_normalized.astype(np.uint8), self.colormap)
-        self.im0 = cv2.addWeighted(self.im0, 0.5, heatmap_colored, 0.5, 0)
+        im0 = cv2.addWeighted(
+            im0,
+            0.5,
+            cv2.applyColorMap(
+                cv2.normalize(self.heatmap, None, 0, 255, cv2.NORM_MINMAX).astype(np.uint8), self.colormap
+            ),
+            0.5,
+            0,
+        )
 
-        if self.env_check and self.view_img:
-            self.display_frames()
+        self.display_output(im0)  # display output with base class function
 
-        return self.im0
-
-    def display_frames(self):
-        """Display frame."""
-        cv2.imshow("Ultralytics Heatmap", self.im0)
-
-        if cv2.waitKey(1) & 0xFF == ord("q"):
-            return
-
-
-if __name__ == "__main__":
-    classes_names = {0: "person", 1: "car"}  # example class names
-    heatmap = Heatmap(classes_names)
+        return im0  # return output image for more usage