--- a/ultralytics/solutions/heatmap.py
+++ b/ultralytics/solutions/heatmap.py
@@ -3,7 +3,7 @@
 import cv2
 import numpy as np
 
-from ultralytics.solutions.object_counter import ObjectCounter
+from ultralytics.solutions.object_counter import ObjectCounter  # Import object counter class
 from ultralytics.utils.plotting import Annotator