--- a/ultralytics/solutions/heatmap.py
+++ b/ultralytics/solutions/heatmap.py
@@ -77,11 +77,14 @@ class Heatmap(ObjectCounter):
 
         # Normalize, apply colormap to heatmap and combine with original image
         im0 = cv2.addWeighted(
-            im0,0.5,
+            im0,
+            0.5,
             cv2.applyColorMap(
                 cv2.normalize(self.heatmap, None, 0, 255, cv2.NORM_MINMAX).astype(np.uint8), self.colormap
             ),
-            0.5, 0,)
+            0.5,
+            0,
+        )
 
         self.display_output(im0)  # display output with base class function
         return im0  # return output image for more usage