--- a/ultralytics/solutions/heatmap.py
+++ b/ultralytics/solutions/heatmap.py
@@ -75,14 +75,18 @@ class Heatmap(ObjectCounter):
         self.display_counts(im0) if self.region is not None else None  # Display the counts on the frame
 
         # Normalize, apply colormap to heatmap and combine with original image
-        im0 = im0 if self.track_data.id is None else cv2.addWeighted(
-            im0,
-            0.5,
-            cv2.applyColorMap(
-                cv2.normalize(self.heatmap, None, 0, 255, cv2.NORM_MINMAX).astype(np.uint8), self.colormap
-            ),
-            0.5,
-            0,
+        im0 = (
+            im0
+            if self.track_data.id is None
+            else cv2.addWeighted(
+                im0,
+                0.5,
+                cv2.applyColorMap(
+                    cv2.normalize(self.heatmap, None, 0, 255, cv2.NORM_MINMAX).astype(np.uint8), self.colormap
+                ),
+                0.5,
+                0,
+            )
         )
 
         self.display_output(im0)  # display output with base class function