… docs/en/guides/heatmaps.md
… tests/test_solutions.py