@glenn-jocher This PR will only update the `heatmap` solution. All other solutions will continue to function as they currently do. I'll open separate PRs for left solutions that need to be optimized :)

### Main changes
- The Model class will be instantiated within the solutions, so users only need to provide the model file path.
- Object tracking will now be handled within the solutions, instead of being managed by an external video processing loop.

### New Code
```python
import cv2
from ultralytics import solutions

cap = cv2.VideoCapture("Path/to/video/file.mp4")
assert cap.isOpened(), "Error reading video file"
w, h, fps = (int(cap.get(x)) for x in (cv2.CAP_PROP_FRAME_WIDTH, cv2.CAP_PROP_FRAME_HEIGHT, cv2.CAP_PROP_FPS))

# Video writer
video_writer = cv2.VideoWriter("heatmap_output.avi", cv2.VideoWriter_fourcc(*"mp4v"), fps, (w, h))

# Init heatmap
heatmap = solutions.Heatmap(
    show=True,
    model="yolo11n.pt",
)

while cap.isOpened():
    success, im0 = cap.read()
    if not success:
        print("Video frame is empty or video processing has been successfully completed.")
        break
    im0 = heatmap.generate_heatmap(im0)
    video_writer.write(im0)

cap.release()
video_writer.release()
cv2.destroyAllWindows()
```

### Output Sample

```CLI
Ultralytics Solutions: ✅ {'model': 'yolo11n.pt', 'region': None, 'line_width': 2, 'show': True, 'show_in': True, 'show_out': True, 'classes': [2], 'up_angle': 145.0, 'down_angle': 90, 'kpts': [6, 8, 10], 'colormap': None}

0: 384x640 (no detections), 178.0ms
Speed: 4.0ms preprocess, 178.0ms inference, 2.0ms postprocess per image at shape (1, 3, 384, 640)
```

## 🛠️ PR Summary

<sub>Made with ❤️ by [Ultralytics Actions](https://github.com/ultralytics/actions)<sub>

### 🌟 Summary
Refactor heatmap generation with streamline improvements and code cleanup.

### 📊 Key Changes
- Removed redundant YOLO model instances for heatmap generation.
- Simplified heatmap class by inheriting from a base class for object counting.
- Enhanced code readability and organization.
- Updated configuration options with more descriptive names.

### 🎯 Purpose & Impact
- **Simplifies Codebase**: Reduces complexity by eliminating unnecessary code, making it easier for developers to understand and maintain.
- **Enhances Performance**: Streamlines heatmap generation, potentially improving execution speed.
- **User-Friendly**: Clearer configuration names improve ease of use for developers and users.
- **Increased Flexibility**: Allows easier modification and extension of heatmap features.