@glenn-jocher 2024-10-06
@RizwanMunawar wow, nice code reduction!!